module7-task3: Добавляет проверку показа откликов
