--- frontend/controllers/Action/AbstractAction.php
@@ -1,14 +1,0 @@
-<?php
-
-
-namespace frontend\controllers\Action;
-
-
-abstract class AbstractAction
-{
-    abstract public function getName();
-
-    abstract public function getRealName();
-
-    abstract public function isRightMethod($user, $task);
-}
