+++ src/controllers/Action/AbstractAction.php
@@ -10,5 +10,5 @@ abstract class AbstractAction
 
     abstract public function getRealName();
 
-    abstract public function isRightMethod(int $user_id, int $owner_id, int $worker_id): bool;
+    abstract public function isRightMethod($user, $task);
 }
